requirements.txt
torch==2.10.0+cpu
torchvision==0.25.0+cpu
numpy==2.0.2
pandas==2.2.2
Pillow==11.3.0
matplotlib==3.10.0
